fix: preserve default permission behavior
