+++ crates/tauri-runtime/src/webview.rs
@@ -4,6 +4,7 @@
 
 //! A layer between raw [`Runtime`] webviews and Tauri.
 //!
+pub use crate::webview_permissions::{PermissionKind, PermissionResponse};
 #[cfg(not(any(target_os = "android", target_os = "ios")))]
 use crate::window::WindowId;
 use crate::{window::is_label_valid, Rect, Runtime, UserEvent};
@@ -72,97 +73,6 @@ pub enum DownloadEvent<'a> {
     /// Indicates if the download succeeded or not.
     success: bool,
   },
-}
-
-/// Permission request response.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
-pub enum PermissionResponse {
-  /// Permission allowed.
-  Allow,
-  /// Permission denied.
-  Deny,
-  /// Use the platform or browser default behavior.
-  #[default]
-  Default,
-}
-
-impl std::fmt::Display for PermissionResponse {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    match self {
-      Self::Allow => write!(f, "allow"),
-      Self::Deny => write!(f, "deny"),
-      Self::Default => write!(f, "default"),
-    }
-  }
-}
-
-/// Permission request kind.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[non_exhaustive]
-pub enum PermissionKind {
-  /// Microphone access permission.
-  Microphone,
-  /// Camera access permission.
-  Camera,
-  /// Geolocation access permission.
-  Geolocation,
-  /// Notifications permission.
-  Notifications,
-  /// Clipboard read permission.
-  ClipboardRead,
-  /// Display capture permission (for getDisplayMedia).
-  DisplayCapture,
-  /// Midi access permission.
-  Midi,
-  /// Sensors (accelerometer, gyroscope, etc.) access permission.
-  Sensors,
-  /// Media key system access permission.
-  MediaKeySystemAccess,
-  /// Local fonts access permission.
-  LocalFonts,
-  /// Window management permission.
-  WindowManagement,
-  /// Pointer lock permission.
-  PointerLock,
-  /// Automatic downloads permission (multiple downloads without user interaction).
-  AutomaticDownloads,
-  /// File system access permission (read/write via File System Access API).
-  ///
-  /// ## Platform-specific
-  /// - **Windows**: Supported via `COREWEBVIEW2_PERMISSION_KIND_FILE_READ_WRITE`.
-  /// - **macOS / Linux / Android / iOS**: Not yet supported by platform backends.
-  FileSystemAccess,
-  /// Media autoplay permission.
-  ///
-  /// ## Platform-specific
-  /// - **Windows**: Supported via `COREWEBVIEW2_PERMISSION_KIND_AUTOPLAY`.
-  /// - **macOS / Linux / Android / iOS**: Not yet supported by platform backends.
-  Autoplay,
-  /// Other unrecognized permission type.
-  Other,
-}
-
-impl std::fmt::Display for PermissionKind {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    match self {
-      Self::Microphone => write!(f, "microphone"),
-      Self::Camera => write!(f, "camera"),
-      Self::Geolocation => write!(f, "geolocation"),
-      Self::Notifications => write!(f, "notifications"),
-      Self::ClipboardRead => write!(f, "clipboard-read"),
-      Self::DisplayCapture => write!(f, "display-capture"),
-      Self::Midi => write!(f, "midi"),
-      Self::Sensors => write!(f, "sensors"),
-      Self::MediaKeySystemAccess => write!(f, "media-key-system-access"),
-      Self::LocalFonts => write!(f, "local-fonts"),
-      Self::WindowManagement => write!(f, "window-management"),
-      Self::PointerLock => write!(f, "pointer-lock"),
-      Self::AutomaticDownloads => write!(f, "automatic-downloads"),
-      Self::FileSystemAccess => write!(f, "file-system-access"),
-      Self::Autoplay => write!(f, "autoplay"),
-      Self::Other => write!(f, "other"),
-    }
-  }
 }
 
 #[cfg(target_os = "android")]
